+++ CHANGELOG.md
@@ -7,13 +7,20 @@ This project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.
 however, insignificant breaking changes does not guarantee a major version bump, see the reasoning [here](https://github.com/kyb3r/modmail/issues/319).
 
 
-# v3.3.1-dev1
+# v3.3.1-dev2
 
 ### Added
 
+- Thread cooldown!
+  - Set via the new config var `thread_cooldown`.
+  - Specify a time for the recipient to wait before allowed to create another thread.
 - "enable" and "disable" support for yes or no config vars.
 - Added "perhaps you meant" section to `?config help`.
 - Multi-command alias is now more stable. With support for a single quote escape `\"`.
+
+### Fixed
+
+- Setting config vars using human time wasn't working.
 
 ### Internal